requirements.txt
fastapi
uvicorn
python-dotenv
praw
tweepy
requests
pydantic
vaderSentiment
transformers
scikit-learn
torch
numpy
pandas
sentencepiece
huggingface-hub
joblib
rich
httpx
cachetools
python-telegram-bot
redis
portia-sdk-python
pytest
pytest-asyncio
gunicorn
